tests/submission/Infraware/Forms/contents/Forms/select_setcustomvalidity.html

<!DOCTYPE HTML>
<html>
 <head>
  <title>Forms</title>
  <script src="../../resources/testharness.js"></script>
  <script src="../../resources/testharnessreport.js"></script>
 </head>
 <body>
	<p>
		<h3>select_setCustomValidity</h3>
	</p>
	
	<hr>

	<div id="log"></div>

	<form method="post"
		  enctype="application/x-www-form-urlencoded"
		  action="" 
		  id="input_form">
		<p>
		 <label for="unittype">Select unit type:</label>
		 <select id="unittype" name="unittype">
		  <option value="1"> Miner </option>
		  <option value="2"> Puffer </option>
		  <option value="3" selected> Snipey </option>
		  <option value="4"> Max </option>
		  <option value="5"> Firebot </option>
		 </select>
		</p>	
	</form>
	<script>

		var select = document.getElementById("unittype");

		try
		{
			select.setCustomValidity("custom error");
			test(function() {
				assert_true(true, "calling of setCustomValidity method is successed.");
			});
		}
		catch (e) {
			test(function() {
				assert_unreached("Error is raised.");
			});
		}

	</script>

 </body>
</html>
